Remove bad advice on module loading
AzureRM is a roll-up module that contaisn 50+ dependent modules.  There is *no need* to load this module in the customer's session.  PowerShell will discover the modules as cmdlets are used.  Adding this step takes an additional minute and a half that si totally unnecessary, and loads 50 modules into memory.  Please remove this as a recommended step.