Clarify Sign In process using service principal
It is confusing the way it is explained how to sing into Azure using the service principal. It is supposed that this login is going to be made in another PowerShell session so variables such as $svcprincipal.ApplicationId won't be available.

Another point is that sometimes using the service principal for login is made to avoid interaction with a user and make an automation task; I am suggesting to add an example of how this could be done.