Update to remove ' MicrosoftDocs/#/1371
The backtick '' character is unnecessary after the pipe "|" character because the pipe acts as a carriage return. For the example on this page, Instead of "Get-AzVM -ResourceGroupName TestGroup | ", please eliminate the backtick character to make it "Get-AzVM -ResourceGroupName TestGroup |". The examples will be more helpful and less confusing for people beginning to use PowerShell. MicrosoftDocs/#/1371