Included Visual Studio PowerShell conflict
There is a conflict that prevents PowerShell 7.x from successfully executing Az module commands if an older version of Visual Studio with Azure PowerShell or the Azure development workload has been installed as this includes the AzureRM module.  This PR draws attention to the possibility of the Visual Studio conflict and provides troubleshooting steps.  This resolves https://github.com/MicrosoftDocs/azure-docs-powershell/issues/1855.