(Maint) Prepare for Hacktoberfest (#2312)
* (MAINT) Add markdownlint rules

This commit adds markdownlint configuration to the repository in sync with
the rules defined in the PowerShell-Docs project. This ensures consistent
application of syntax rules across the content.

* (MAINT) Add cSpell to repository

This change introduces cSpell to the repository for managing
spelling across the files. In this initial draft, it adds the
upstream dictionary from MicrosoftDocs/PowerShell-Docs and
scaffolds a nearly-empty dictionary for the Azure PowerShell
modules to make it easier to define terms in the future.

It also removes the directive in gitignore so that the .vscode folder
will be tracked.

* (MAINT) Add extension suggestions

This change adds extension suggestions to the repository to ease contributions
by providing a list of helpful extensions for writing and maintaining the
documentation.

* (MAINT) Add issue templates

This change introduces GitHub issue templates and the issue picker to the
project, enabling an improved UX for people filing issues against the
repository. It also includes the issue template for Hacktoberfest 2022,
to make participation easier.

The configuration for the article issue template is minimal, lacking version
or module selection, but those fields can be added in the future if required.

* (MAINT) Add Repo automation

This change introduces some repository automation from the PowerShell-Docs
repositories.

In summary:

1. The CODEOWNERS file automatically adds the repository owner(s) as reviewers
   to incoming PRs.
2. The Pull Request template ensures a standard UX and process for incoming PRs.
3. The checklist workflow uses the PR template to verify that non-admin
   contributors are following the processes.
4. The expectations workflow uses a static message (defined in the /expectations.md
   file in the messages folder) to welcome external contributions and set expectations
   around process for their PR.
5. The dependabot configuration verifies that actions are up to date.